core buffer

+++ LuaSTG/Core/Graphics/Device.hpp
@@ -129,7 +129,7 @@ namespace Core::Graphics
 	};
 
 	struct IBuffer : IObject {
-		virtual bool map(size_t size_in_bytes, bool discard, void** out_pointer) = 0;
+		virtual bool map(uint32_t size_in_bytes, bool discard, void** out_pointer) = 0;
 		virtual bool unmap() = 0;
 	};
 
@@ -149,8 +149,8 @@ namespace Core::Graphics
 		virtual void* getNativeHandle() = 0;
 		virtual void* getNativeRendererHandle() = 0;
 
-		virtual bool createVertexBuffer(size_t size_in_bytes, IBuffer** output) = 0;
-		virtual bool createIndexBuffer(size_t size_in_bytes, IBuffer** output) = 0;
+		virtual bool createVertexBuffer(uint32_t size_in_bytes, IBuffer** output) = 0;
+		virtual bool createIndexBuffer(uint32_t size_in_bytes, IBuffer** output) = 0;
 
 		virtual bool createTextureFromFile(StringView path, bool mipmap, ITexture2D** pp_texture) = 0;
 		//virtual bool createTextureFromMemory(void const* data, size_t size, bool mipmap, ITexture2D** pp_texture) = 0;
